// src/generated/models/Poce_assignmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Poce_assignmentspoce_status = {
  413450000: 'Active',
  413450001: 'Completed'
} as const;
export type Poce_assignmentspoce_status = keyof typeof Poce_assignmentspoce_status;
export const Poce_assignmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Poce_assignmentsstatecode = keyof typeof Poce_assignmentsstatecode;
export const Poce_assignmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Poce_assignmentsstatuscode = keyof typeof Poce_assignmentsstatuscode;

export interface Poce_assignmentsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  poce_assignmentid: string;
  poce_createdon?: string;
  poce_department?: string;
  poce_executivename?: string;
  poce_expectedreturndate: string;
  poce_modifiedon?: string;
  poce_purpose?: string;
  poce_startdate: string;
  poce_status: Poce_assignmentspoce_status;
  "poce_Vehicle@odata.bind": string;
  statecode: Poce_assignmentsstatecode;
  statuscode?: Poce_assignmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Poce_assignments extends Poce_assignmentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  poce_statusname?: string;
  poce_vehiclename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  poce_vehicle?: object;
  _poce_vehicle_value?: string;
}